Added an example Auth setup for the browser
Added auth project that shows how to securely auth to a webrelay from the browser. I thought it would be helpful for users looking to develop their own browser client.

This example project will be expanded in the future to include browser messaging, order handling, and more. The goal is to eventually develop a full browser node for OB.